backend/requirements.txt
azure-functions==1.21.0
azure-cosmos==4.7.0
azure-storage-blob==12.19.1
azure-identity==1.16.1
azure-messaging-webpubsubservice==1.1.0
openai==1.30.0
PyJWT==2.8.0
requests==2.31.0
cryptography==42.0.7
